Add Dockerfile
Uses the latest master build of Flink.

This does not work on DC/OS yet as JAVA_HOME is required to be set on
the task managers which DC/OS does not do by default (this behavior will
not be changed). The new coming feature allows to deploy a task manager
with a container image which would contain in this case a working java
version.