Update Flink
As the Flink AppMaster is now started via the appmaster script, this
changes the game. The script accepts configuration variables and allows
to merge settings cleanly.

For that reason the docker image is made more "atomic", it grabs a
default configuration (flink-conf.yaml) and can run standalone without
extra arguments.